release(group, review):

See merge request s12-fintech-finance-sub1/S12P21C208!67

+++ backend/src/main/java/com/patriot/fourlipsclover/notification/service/NotificationService.java
@@ -20,7 +20,7 @@ public class NotificationService {
      * 그룹 가입 요청 알림 전송
      * 가입 요청이 발생하면, 그룹 리더(Member)를 대상으로 알림 엔티티를 생성하여 DB에 저장합니다.
      */
-    public void sendGroupJoinRequestNotification(Integer groupId, Integer groupLeaderId) {
+    public void sendGroupJoinRequestNotification(Integer groupId, Long groupLeaderId) {
         // 그룹 리더(Member) 엔티티를 조회합니다.
         Member groupLeader = memberRepository.findById(groupLeaderId)
                 .orElseThrow(() -> new RuntimeException("그룹 리더 회원을 찾을 수 없습니다. id=" + groupLeaderId));
